fix: Clean up clippy warnings and improve code quality for v0.8.2
- Applied clippy auto-fixes across workspace
- Fixed empty line after doc comments
- Replaced manual strip_prefix with idiomatic method calls
- Fixed if_same_then_else warnings
- Fixed IncrementalParser import in examples
- Removed unnecessary mut qualifiers
- Improved code idioms throughout

+++ crates/perl-parser/tests/lsp_concurrency.rs
@@ -672,7 +672,7 @@ fn test_semantic_tokens_consistency() {
                     "params": {
                         "textDocument": {
                             "uri": uri,
-                            "version": id as i32 + 1
+                            "version": id + 1
                         },
                         "contentChanges": [{
                             "text": format!("package Test;\nuse strict;\nmy $var = {};\nsub func {{ return $var; }}", id * 10)
